learning_voice_agent/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
anthropic==0.18.1
openai==1.10.0
twilio==8.11.0
redis==5.0.1
aioredis==2.0.1
aiosqlite==0.19.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.3
httpx==0.26.0
jinja2==3.1.3